ci/requirements.txt
linkify-it-py==2.1.0 ; python_version >= "3.13"
markdown-it-py==4.0.0 ; python_version >= "3.13"
mdurl==0.1.2 ; python_version >= "3.13"
python-dotenv==1.2.2 ; python_version >= "3.13"
uc-micro-py==2.0.0 ; python_version >= "3.13"
